+++ Cookie.API/Protocol/Network/Messages/Connection/CredentialsAcknowledgementMessage.cs
@@ -4,18 +4,16 @@ namespace Cookie.API.Protocol.Network.Messages.Connection
 {
     public class CredentialsAcknowledgementMessage : NetworkMessage
     {
-        public const uint ProtocolId = 6314;
+        public const ushort ProtocolId = 6314;
 
-        public override uint MessageID => ProtocolId;
+        public override ushort MessageID => ProtocolId;
 
         public override void Serialize(IDataWriter writer)
         {
-            //
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            //
         }
     }
 }
