+++ Cookie.API/Protocol/Network/Messages/Game/Approach/AuthenticationTicketAcceptedMessage.cs
@@ -4,18 +4,16 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Approach
 {
     public class AuthenticationTicketAcceptedMessage : NetworkMessage
     {
-        public const uint ProtocolId = 111;
+        public const ushort ProtocolId = 111;
 
-        public override uint MessageID => ProtocolId;
+        public override ushort MessageID => ProtocolId;
 
         public override void Serialize(IDataWriter writer)
         {
-            //
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            //
         }
     }
 }
