+++ Cookie.API/Protocol/Network/Messages/Game/Friend/IgnoredGetListMessage.cs
@@ -4,18 +4,16 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Friend
 {
     public class IgnoredGetListMessage : NetworkMessage
     {
-        public const uint ProtocolId = 5676;
+        public const ushort ProtocolId = 5676;
 
-        public override uint MessageID => ProtocolId;
+        public override ushort MessageID => ProtocolId;
 
         public override void Serialize(IDataWriter writer)
         {
-            //
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            //
         }
     }
 }
