+++ Cookie.API/Protocol/Network/Messages/Game/Friend/SpouseGetInformationsMessage.cs
@@ -4,18 +4,16 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Friend
 {
     public class SpouseGetInformationsMessage : NetworkMessage
     {
-        public const uint ProtocolId = 6355;
+        public const ushort ProtocolId = 6355;
 
-        public override uint MessageID => ProtocolId;
+        public override ushort MessageID => ProtocolId;
 
         public override void Serialize(IDataWriter writer)
         {
-            //
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            //
         }
     }
 }
