+++ Cookie.API/Protocol/Network/Messages/Game/Initialization/CharacterLoadingCompleteMessage.cs
@@ -4,18 +4,16 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Initialization
 {
     public class CharacterLoadingCompleteMessage : NetworkMessage
     {
-        public const uint ProtocolId = 6471;
+        public const ushort ProtocolId = 6471;
 
-        public override uint MessageID => ProtocolId;
+        public override ushort MessageID => ProtocolId;
 
         public override void Serialize(IDataWriter writer)
         {
-            //
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            //
         }
     }
 }
